+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Breaking change: Split tooltip component into a component `Tooltip` and hoc:
   `withTooltip`.
   To be retro-compatible, you should replace `Tooltip` component with `InfoWithTooltip`.
+- Feature: Add new `InfoLines` component.
 - Fix: Add `width` attribute on payment button svg tag (fixes firefox bug).
 - Feature: Add `color-error` color text utility.
 - Feature: For `xs` media-query, `FormActions` component is placed in columns.